Move event code to server.py
